pronavi/requirements.txt
flask>=3.0
pymssql>=2.3
openpyxl>=3.1
python-dotenv>=1.0
